Refactor GAME_COMMAND_REMOVE_TRACK to game action.
