+++ src/openrct2/ride/track.c
@@ -763,7 +763,7 @@ static bool track_add_station_element(sint32 x, sint32 y, sint32 z, sint32 direc
  *
  *  rct2: 0x006C494B
  */
-static bool track_remove_station_element(sint32 x, sint32 y, sint32 z, sint32 direction, sint32 rideIndex, sint32 flags)
+bool track_remove_station_element(sint32 x, sint32 y, sint32 z, sint32 direction, sint32 rideIndex, sint32 flags)
 {
     sint32 removeX = x;
     sint32 removeY = y;
@@ -1437,292 +1437,6 @@ void game_command_place_track(sint32 *eax, sint32 *ebx, sint32 *ecx, sint32 *edx
         );
 }
 
-static money32 track_remove(uint8 type, uint8 sequence, sint16 originX, sint16 originY, sint16 originZ, uint8 rotation, uint8 flags){
-    gCommandExpenditureType = RCT_EXPENDITURE_TYPE_RIDE_CONSTRUCTION;
-    gCommandPosition.x = originX + 16;
-    gCommandPosition.y = originY + 16;
-    gCommandPosition.z = originZ;
-    sint16 trackpieceZ = originZ;
-
-    switch (type){
-    case TRACK_ELEM_BEGIN_STATION:
-    case TRACK_ELEM_MIDDLE_STATION:
-        type = TRACK_ELEM_END_STATION;
-        break;
-    }
-
-    if (!(flags & (1 << 3)) && game_is_paused() && !gCheatsBuildInPauseMode){
-        gGameCommandErrorText = STR_CONSTRUCTION_NOT_POSSIBLE_WHILE_GAME_IS_PAUSED;
-        return MONEY32_UNDEFINED;
-    }
-
-    uint8 found = 0;
-    rct_map_element* mapElement = map_get_first_element_at(originX / 32, originY / 32);
-    if (mapElement == NULL)
-    {
-        log_warning("Invalid coordinates for track removal. x = %d, y = %d", originX, originY);
-        return MONEY32_UNDEFINED;
-    }
-    do{
-        if (mapElement->base_height * 8 != originZ)
-            continue;
-
-        if (map_element_get_type(mapElement) != MAP_ELEMENT_TYPE_TRACK)
-            continue;
-
-        if ((map_element_get_direction(mapElement)) != rotation)
-            continue;
-
-        if (map_element_get_track_sequence(mapElement) != sequence)
-            continue;
-
-        // Probably should add a check for ghost here as well!
-
-        uint8 track_type = mapElement->properties.track.type;
-        switch (track_type){
-        case TRACK_ELEM_BEGIN_STATION:
-        case TRACK_ELEM_MIDDLE_STATION:
-            track_type = TRACK_ELEM_END_STATION;
-            break;
-        }
-
-        if (track_type != type)
-            continue;
-
-        found = 1;
-        break;
-    } while (!map_element_is_last_for_tile(mapElement++));
-
-    if (!found){
-        return MONEY32_UNDEFINED;
-    }
-
-    if (mapElement->flags & MAP_ELEMENT_FLAG_INDESTRUCTIBLE_TRACK_PIECE){
-        gGameCommandErrorText = STR_YOU_ARE_NOT_ALLOWED_TO_REMOVE_THIS_SECTION;
-        return MONEY32_UNDEFINED;
-    }
-
-    uint8 rideIndex = mapElement->properties.track.ride_index;
-    type = mapElement->properties.track.type;
-    bool isLiftHill = track_element_is_lift_hill(mapElement);
-
-    Ride* ride = get_ride(rideIndex);
-    const rct_preview_track* trackBlock = get_track_def_from_ride(ride, type);
-    trackBlock += map_element_get_track_sequence(mapElement);
-
-    uint8 originDirection = map_element_get_direction(mapElement);
-    switch (originDirection){
-    case 0:
-        originX -= trackBlock->x;
-        originY -= trackBlock->y;
-        break;
-    case 1:
-        originX -= trackBlock->y;
-        originY += trackBlock->x;
-        break;
-    case 2:
-        originX += trackBlock->x;
-        originY += trackBlock->y;
-        break;
-    case 3:
-        originX += trackBlock->y;
-        originY -= trackBlock->x;
-        break;
-    }
-
-    originZ -= trackBlock->z;
-
-    money32 cost = 0;
-
-    trackBlock = get_track_def_from_ride(ride, type);
-    for (; trackBlock->index != 255; trackBlock++){
-        sint16 x = originX, y = originY, z = originZ;
-
-        switch (originDirection){
-        case 0:
-            x += trackBlock->x;
-            y += trackBlock->y;
-            break;
-        case 1:
-            x += trackBlock->y;
-            y -= trackBlock->x;
-            break;
-        case 2:
-            x -= trackBlock->x;
-            y -= trackBlock->y;
-            break;
-        case 3:
-            x -= trackBlock->y;
-            y += trackBlock->x;
-            break;
-        }
-
-        z += trackBlock->z;
-
-        map_invalidate_tile_full(x, y);
-
-        trackpieceZ = z;
-
-        found = 0;
-        mapElement = map_get_first_element_at(x / 32, y / 32);
-        do{
-            if (mapElement == NULL)
-                break;
-
-            if (mapElement->base_height != z / 8)
-                continue;
-
-            if (map_element_get_type(mapElement) != MAP_ELEMENT_TYPE_TRACK)
-                continue;
-
-            if ((map_element_get_direction(mapElement)) != rotation)
-                continue;
-
-            if (map_element_get_track_sequence(mapElement) != trackBlock->index)
-                continue;
-
-            if (mapElement->properties.track.type != type)
-                continue;
-
-            found = 1;
-            break;
-        } while (!map_element_is_last_for_tile(mapElement++));
-
-        if (!found){
-            log_error("Track map element part not found!");
-            return MONEY32_UNDEFINED;
-        }
-
-        sint32 entranceDirections;
-        if (ride_type_has_flag(ride->type, RIDE_TYPE_FLAG_FLAT_RIDE)) {
-            entranceDirections = FlatRideTrackSequenceProperties[type][0];
-        }
-        else {
-            entranceDirections = TrackSequenceProperties[type][0];
-        }
-
-        if (entranceDirections & TRACK_SEQUENCE_FLAG_ORIGIN && (map_element_get_track_sequence(mapElement) == 0)){
-            if (!track_remove_station_element(x, y, z / 8, rotation, rideIndex, 0)) {
-                return MONEY32_UNDEFINED;
-            }
-        }
-
-        rct_map_element* surfaceElement = map_get_surface_element_at(x / 32, y / 32);
-        if (surfaceElement == NULL){
-            return MONEY32_UNDEFINED;
-        }
-
-        sint8 _support_height = mapElement->base_height - surfaceElement->base_height;
-        if (_support_height < 0){
-            _support_height = 10;
-        }
-
-        cost += (_support_height / 2) * RideTrackCosts[ride->type].support_price;
-
-        if (!(flags & GAME_COMMAND_FLAG_APPLY))
-            continue;
-
-        if (entranceDirections & (1 << 4) && (map_element_get_track_sequence(mapElement) == 0)){
-            if (!track_remove_station_element(x, y, z / 8, rotation, rideIndex, GAME_COMMAND_FLAG_APPLY)) {
-                return MONEY32_UNDEFINED;
-            }
-        }
-
-        if (ride_type_has_flag(ride->type, RIDE_TYPE_FLAG_TRACK_MUST_BE_ON_WATER)){
-            surfaceElement->type &= ~(1 << 6);
-        }
-
-        invalidate_test_results(rideIndex);
-        footpath_queue_chain_reset();
-        if (!gCheatsDisableClearanceChecks || !(mapElement->flags & MAP_ELEMENT_FLAG_GHOST)) {
-            footpath_remove_edges_at(x, y, mapElement);
-        }
-        map_element_remove(mapElement);
-        if (!(flags & GAME_COMMAND_FLAG_GHOST)){
-            sub_6CB945(rideIndex);
-            ride_update_max_vehicles(rideIndex);
-        }
-    }
-
-    if (flags & GAME_COMMAND_FLAG_APPLY){
-        switch (type){
-        case TRACK_ELEM_ON_RIDE_PHOTO:
-            ride->lifecycle_flags &= ~RIDE_LIFECYCLE_ON_RIDE_PHOTO;
-            break;
-        case TRACK_ELEM_CABLE_LIFT_HILL:
-            ride->lifecycle_flags &= ~RIDE_LIFECYCLE_CABLE_LIFT_HILL_COMPONENT_USED;
-            break;
-        case TRACK_ELEM_BLOCK_BRAKES:
-            ride->num_block_brakes--;
-            if (ride->num_block_brakes == 0){
-                ride->window_invalidate_flags |= RIDE_INVALIDATE_RIDE_OPERATING;
-                ride->mode = RIDE_MODE_CONTINUOUS_CIRCUIT;
-                if (ride->type == RIDE_TYPE_LIM_LAUNCHED_ROLLER_COASTER){
-                    ride->mode = RIDE_MODE_POWERED_LAUNCH;
-                }
-            }
-            break;
-        }
-
-        switch (type){
-        case TRACK_ELEM_25_DEG_UP_TO_FLAT:
-        case TRACK_ELEM_60_DEG_UP_TO_FLAT:
-        case TRACK_ELEM_DIAG_25_DEG_UP_TO_FLAT:
-        case TRACK_ELEM_DIAG_60_DEG_UP_TO_FLAT:
-            if (!isLiftHill)
-                break;
-            // Fall through
-        case TRACK_ELEM_CABLE_LIFT_HILL:
-            ride->num_block_brakes--;
-            break;
-        }
-    }
-
-    money32 price = RideTrackCosts[ride->type].track_price;
-    if (ride_type_has_flag(ride->type, RIDE_TYPE_FLAG_FLAT_RIDE)) {
-        price *= FlatRideTrackPricing[type];
-    }
-    else {
-        price *= TrackPricing[type];
-    }
-    price >>= 16;
-    price = (price + cost) / 2;
-    if (ride->lifecycle_flags & RIDE_LIFECYCLE_EVER_BEEN_OPENED)
-        price *= -7;
-    else
-        price *= -10;
-
-    if (gGameCommandNestLevel == 1) {
-        rct_xyz16 coord;
-        coord.x = originX + 16;
-        coord.y = originY + 16;
-        coord.z = trackpieceZ;
-        network_set_player_last_action_coord(network_get_player_index(game_command_playerid), coord);
-    }
-
-    if (gParkFlags & PARK_FLAGS_NO_MONEY)
-        return 0;
-    else
-        return price;
-}
-
-/**
- *
- *  rct2: 0x006C5B69
- */
-void game_command_remove_track(sint32 *eax, sint32 *ebx, sint32 *ecx, sint32 *edx, sint32 *esi, sint32 *edi, sint32 *ebp)
-{
-    *ebx = track_remove(
-        *edx & 0xFF,
-        (*edx >> 8) & 0xFF,
-        *eax & 0xFFFF,
-        *ecx & 0xFFFF,
-        *edi & 0xFFFF,
-        (*ebx >> 8) & 0xFF,
-        *ebx & 0xFF
-        );
-    return;
-}
 
 static uint8 maze_element_get_segment_bit(uint16 x, uint16 y) {
     uint8 minorX = x & 0x1F; // 0 or 16
